requirements.txt
certifi==2020.6.20
chardet==3.0.4
croniter==0.3.34
idna==2.10
natsort==7.0.1
oauthlib==3.1.0
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.8.0
urllib3==1.25.9
